% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smuggler.R
\name{meta_to_string}
\alias{meta_to_string}
\title{get a meta data string for an object}
\usage{
meta_to_string(fameinfo, objnam)
}
\arguments{
\item{fameinfo}{a list containing fame object meta data}

\item{objnam}{object name}
}
\value{
string containing meta data
}
\description{
get a meta data string for an object
}
