% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smuggler.R
\name{newEntry}
\alias{newEntry}
\title{Construct a List entry with FAME data and metadata}
\usage{
newEntry(data, desc = NULL, docu = NULL, class = rhli::HSERIE,
  range = NULL, type = rhli::HPRECN, basis = rhli::HBSBUS,
  obse = rhli::HOBEND)
}
\arguments{
\item{data}{data value(s) to store}

\item{desc}{description}

\item{docu}{documentation}

\item{class}{object class HLI code}

\item{range}{FAME range of object data (if series)}

\item{type}{object type HLI code}

\item{basis}{object basis HLI code}

\item{obse}{object observed attribute}
}
\value{
FAME database object (data and metadata as nested R list)
}
\description{
Construct a List entry with FAME data and metadata
}
\examples{
entry <- newEntry('String contents',class=rhli::HSCALA,type=rhli::HSTRNG)
    
}
