% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smuggler.R
\name{print_catalog}
\alias{print_catalog}
\title{Print a catalog}
\usage{
print_catalog(famedata, list.len = -1)
}
\arguments{
\item{famedata}{List containing FAME data to display}

\item{list.len}{maximum entries to display}
}
\description{
Print a catalog
}
\examples{
myStuff <- List()
entry <- newEntry('String contents',class=rhli::HSCALA,type=rhli::HSTRNG)
myStuff$put('str',entry)
print_catalog(myStuff)

}
