% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smuggler.R
\name{put}
\alias{put}
\title{put data for FAME in a list}
\usage{
put(mylist, objectname, data, desc = NULL, docu = NULL,
  class = rhli::HSERIE, range = NULL, type = rhli::HPRECN,
  basis = rhli::HBSBUS, obse = rhli::HOBEND)
}
\arguments{
\item{mylist}{target qoma.smuggler::List for put() operation}

\item{objectname}{object name}

\item{data}{data value(s) to store}

\item{desc}{description}

\item{docu}{documentation}

\item{class}{object class HLI code}

\item{range}{FAME range of object data (if series)}

\item{type}{object type HLI code}

\item{basis}{object basis HLI code}

\item{obse}{object observed attribute}
}
\value{
success TRUE or FALSE
}
\description{
put data for FAME in a list
}
