% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qomautilr.R
\name{to_fame_range}
\alias{to_fame_range}
\title{Construct a date range}
\usage{
to_fame_range(freq_, start_str_, end_str_)
}
\arguments{
\item{freq}{FAME HLI frequency constant}

\item{start_str}{start date literal}

\item{end_str}{end date literal}
}
\value{
date range
}
\description{
Construct a date range
}
