% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smuggler.R
\name{to_lubridate_date}
\alias{to_lubridate_date}
\title{create a lubridate date}
\usage{
to_lubridate_date(fame_freq, date)
}
\arguments{
\item{fame_freq}{FAME frequency code}

\item{date}{FAME date}
}
\value{
lubridate date
}
\description{
create a lubridate date
}
