% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smuggler.R
\name{to_lubridate_index}
\alias{to_lubridate_index}
\title{create a lubridate index}
\usage{
to_lubridate_index(rng)
}
\arguments{
\item{rng}{FAME range}
}
\value{
tibble with lubridate date column
}
\description{
create a lubridate index
}
