% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smuggler.R
\name{type_to_string}
\alias{type_to_string}
\title{convert a fame type code to a string}
\usage{
type_to_string(type_code_)
}
\arguments{
\item{type_code_}{FAME HLI type code}
}
\value{
type string
}
\description{
convert a fame type code to a string
}
