% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qomautilr.R
\docType{class}
\name{List-class}
\alias{List-class}
\alias{List}
\title{Mutable list}
\description{
Mutable list
}
\section{Methods}{

\describe{
\item{\code{get(objnam = NULL)}}{Get an element of the list}

\item{\code{get_data(objnam = NULL)}}{Get data}

\item{\code{get_meta(objnam = NULL)}}{Get meta data}

\item{\code{initialize(l0 = list())}}{Initialize a list.}
}}

