% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smuggler.R
\name{meta_to_string}
\alias{meta_to_string}
\title{Get a meta data string for an object}
\usage{
meta_to_string(fameMeta, objnam)
}
\arguments{
\item{fameMeta}{a list containing fame object meta data}

\item{objnam}{object name}
}
\value{
string containing meta data
}
\description{
Get a meta data string for an object
}
\examples{
dbname <- file.path(Sys.getenv("FAME"),"util","driecon")
famedb <- read_fame(dbname)
meta <- famedb$get_meta('GDP')
cat(meta_to_string(famedb$get('GDP')$meta, objnam))
}
