% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qomautilr.R
\name{meta_to_string}
\alias{meta_to_string}
\title{get a meta data string for an object}
\usage{
meta_to_string(fameinfo, objnam)
}
\arguments{
\item{objnam}{object name}

\item{famedata}{a list containing fame database objects}
}
\value{
string containing meta data
}
\description{
get a meta data string for an object
}
