% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qomautilr.R
\name{put}
\alias{put}
\title{put data for FAME in a list}
\usage{
put(mylist, objectname, data, desc = NULL, docu = NULL,
  class = rhli::HSERIE, range = NULL, type = rhli::HPRECN,
  basis = rhli::HBSBUS, obse = rhli::HOBEND)
}
\arguments{
\item{objectname}{object name}

\item{data}{data value(s) to store}

\item{desc}{description}

\item{docu}{documentation}

\item{basis}{object basis attribute}

\item{obse}{object observed attribute}

\item{otype}{object type attribute}
}
\value{
success TRUE or FALSE
}
\description{
put data for FAME in a list
}
