% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smuggler.R
\name{read_fame}
\alias{read_fame}
\title{Read a FAME database into an R list}
\usage{
read_fame(dbname_, wilnam_ = "?", fame_range_ = NULL)
}
\arguments{
\item{dbname_}{FAME database filename}

\item{wilnam_}{object name wildcard}

\item{fame_range_}{FAME range to limit data retrieval}
}
\description{
Read a FAME database into an R list
}
