% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smuggler.R
\name{to_fame_range}
\alias{to_fame_range}
\title{Construct a date range}
\usage{
to_fame_range(freq_, start_str_, end_str_)
}
\arguments{
\item{freq_}{FAME frequency HLI code}

\item{start_str_}{FAME start date literal (e.g. "2018M12")}

\item{end_str_}{FAME end date literal}
}
\value{
date range
}
\description{
Construct a date range
}
