% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smuggler.R
\name{write_fame}
\alias{write_fame}
\title{Write FAME db}
\usage{
write_fame(dbname_, container)
}
\arguments{
\item{dbname_}{FAME database filename}

\item{container}{List with data to write}
}
\description{
Write FAME db
}
\examples{
mydb <- List()
# construct an entry for FAME scalar string
entry <- newEntry('String contents',class=rhli::HSCALA,type=rhli::HSTRNG)
mydb$put('str',entry)
dbfile <- file.path(tempdir(),'tmp.db')
write_fame(dbfile,mydb)

}
