% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smuggler.R
\name{write_fame}
\alias{write_fame}
\title{wrte fame db}
\usage{
write_fame(dbname_, container)
}
\arguments{
\item{dbname_}{FAME database filename}

\item{container}{List with data to write}
}
\description{
wrte fame db
}
